# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_joaofilippe_subclub_internal_web_common.Response:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  internal_web_plan.PaginatedPlanResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_web_plan.PlanDTO'
        type: array
      totalCount:
        type: integer
    type: object
  internal_web_plan.PlanDTO:
    properties:
      active:
        type: boolean
      code:
        type: string
      createdAt:
        type: string
      description:
        type: string
      discountValue:
        type: number
      id:
        type: string
      imageUrl:
        type: string
      intervalDays:
        type: integer
      name:
        type: string
      price:
        type: number
      productValue:
        type: number
    type: object
  internal_web_plan.PlanInputDTO:
    properties:
      active:
        type: boolean
      code:
        type: string
      description:
        type: string
      discountValue:
        type: number
      imageUrl:
        type: string
      intervalDays:
        type: integer
      name:
        type: string
      price:
        type: number
      productValue:
        type: number
    type: object
  internal_web_product.PaginatedProductResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_web_product.ProductDTO'
        type: array
      totalCount:
        type: integer
    type: object
  internal_web_product.ProductDTO:
    properties:
      active:
        type: boolean
      category:
        type: string
      code:
        type: string
      costPrice:
        type: number
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      imageUrl:
        type: string
      name:
        type: string
    type: object
  internal_web_product.ProductInputDTO:
    properties:
      active:
        type: boolean
      category:
        type: string
      code:
        type: string
      costPrice:
        type: number
      description:
        type: string
      imageUrl:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend application for SubClub
  title: SubClub API
  version: "1.0"
paths:
...
  /api/v1/plans:
    get:
      description: Get a paginated list of plans
      parameters:
      - description: Search by name or code
        in: query
        name: search
        type: string
      - description: Filter by active status
        in: query
        name: active
        type: boolean
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_plan.PaginatedPlanResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
      summary: List plans
      tags:
      - plans
    post:
      consumes:
      - application/json
      description: Creates a new subscription plan
      parameters:
      - description: Plan data
        in: body
        name: plan
        required: true
        schema:
          $ref: '#/definitions/internal_web_plan.PlanInputDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_web_plan.PlanDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
      summary: Create a plan
      tags:
      - plans
  /api/v1/plans/{id}:
    delete:
      description: Soft deletes a plan by ID
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
      summary: Delete a plan
      tags:
      - plans
    get:
      description: Retrieves a plan by its UUID
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_plan.PlanDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
      summary: Get a plan by ID
      tags:
      - plans
    put:
      consumes:
      - application/json
      description: Updates an existing plan
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated plan data
        in: body
        name: plan
        required: true
        schema:
          $ref: '#/definitions/internal_web_plan.PlanInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_plan.PlanDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
      summary: Update a plan
      tags:
      - plans
  /api/v1/products:
    get:
      description: Get a paginated list of products
      parameters:
      - description: Search by name or code
        in: query
        name: search
        type: string
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Filter by active status
        in: query
        name: active
        type: boolean
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_product.PaginatedProductResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
      summary: List products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Creates a new product
      parameters:
      - description: Product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/internal_web_product.ProductInputDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_web_product.ProductDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
      summary: Create a product
      tags:
      - products
  /api/v1/products/{id}:
    delete:
      description: Soft deletes a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
      summary: Delete a product
      tags:
      - products
    get:
      description: Retrieves a product by its UUID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_product.ProductDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
      summary: Get a product by ID
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Updates an existing product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/internal_web_product.ProductInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_product.ProductDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
      summary: Update a product
      tags:
      - products
...
swagger: "2.0"